create Module Controller
