+++ src/ts-framework/Controller/ActionDecorator.ts
@@ -1,3 +1,22 @@
+import { Controller } from "./Controller";
+
 /**
- * Created by jhon on 27/04/16.
+ * Action decorator
+ * @param parameters
+ * @returns {TypedPropertyDescriptor<any>}
+ * @decorator
  */
+export function action(parameters: Object = null)
+{
+    return function (target:Controller, propertyKey:string, descriptor:TypedPropertyDescriptor<any>) {
+        if (parameters) {
+            var aux = {};
+            if (target.decorate)
+                aux = target.decorate;
+            aux[propertyKey] = parameters;
+            target.decorate = aux;
+        }
+
+        return descriptor;
+    };
+}
