+++ src/ts-framework/Controller/DataModelController.ts
@@ -1,3 +1,44 @@
+import { Controller } from "./Controller";
+
 /**
- * Created by jhon on 27/04/16.
+ * TS-Framework application
+ * DataModelController.ts - Register methods for communication with ORM
+ * extends Controller
  */
+export class DataModelController extends Controller
+{
+    /**
+     * Create new Object
+     */
+    public create()
+    {
+        // ...
+    }
+
+    /**
+     * Update Object specified with ID
+     * @param {string} id
+     */
+    public update(id: string)
+    {
+        // ...
+    }
+
+    /**
+     * Delete Object specified with ID
+     * @param {string} id
+     */
+    public delete(id: string)
+    {
+        // ...
+    }
+
+    /**
+     * Find object by id
+     * @param {string} id
+     */
+    public find(id: string)
+    {
+        // ...
+    }
+}
